… regression-test/data/nereids_tpcds_shape_sf100_p0/no_stats_shape/query64.out